Moved detail level code related to payment results into PaymentRequestDialog widget
